+++ src/models/Message.ts
@@ -1,0 +1,9 @@
+export type Messages = Message[];
+
+export interface Message {
+  authorName: string;
+  authorId: string;
+  text: string;
+  createdAt: Object;
+  key: string;
+}
