+++ src/models/Room.ts
@@ -1,0 +1,5 @@
+export interface RoomInfo {
+  masterId: string;
+  roomId: string;
+  roomName: string;
+}
